+++ content/파일 압축하기/index.mdx
@@ -6,7 +6,7 @@ tags: ['Browser', 'Performance']
 hero_image: './파일_압축하기_히어로_이미지.jpeg'
 hero_image_alt: '파일 압축하기 히어로 이미지'
 ---
-import { MDXImageWrapper, ToggleList } from 'components';
+import { MDXImageWrapper } from 'components';
 
 # 들어가며
